+++ ChangeLog.md
@@ -5,9 +5,11 @@
 - Implemented "Jet Fuel" tier upgrade which increases the acceleration of the jetpack while in flight.
 - Implemented "Midas Touch" tier upgrade which increases the value of scrap found in the moons.
 - Implemented "Quick Hands" tier upgrade which increases the player's interaction speed.
+- Added configuration to always show items associated to upgrades when Item Progression Mode is toggled on.
 - Changed upgrade display to show the price number red if the players do not have enough credits to purchase it.
 - Changed upgrade store to show first pick between shared and individual upgrades if any are configured as so and show the list of upgrades when selected.
   - If all enabled upgrades are either individual or shared, it will do the same behaviour as before.
+- Fixed issue with Item Progression Mode selecting items that are not considered scrap.
 
 </details>